requirements.txt
requests==3.11.9
beautifulsoup4
streamlit==1.33.0
numpy
Pillow
flask
keras~=2.15.0
tensorflow~=2.15.0